reorganize the package ast source files (#5370)
This commit merges the ast/zed package into compiler/ast and simplifies
the ast layout.  It was good to keep them separate till now but
now that the dust has settled on the super-structured data model,
it all belongs together and hangs together nicely.

This change is needed to break an import sycle in a subsequent PR that
will simplify the ast.Node implementations, which is in turn needed,
by the SuperSQL additions.

We also moved the dag data structure from compiler/ast/dag to
compiler/dag as the dag describes the physical runtime query
execution structure and has nothing to do with syntax of the
query language.

Note that there is no new logic introduced here.  This is strictly
source code file movement and name changes.