unify name-y things in grammar (#5373)
This commit unifies the various types of named things in the PEG grammar
to a single pattern called ast.Name and gets rid of ast.String and
ast.QuotedString.  Now that the locators are all computed from the
underlying rule, we can easily have a single ast.Name that represents
different variations of names.  These changes will be used by a subequent PR
that will unify the get/file/pool operators into a single "from", which
is in turn needed by the upcoming SuperSQL additions.