clarify PEG grammar around lean-forward idioms of SuperPipe (#5380)
This commit reorganizes the PEG grammar a bit to put all the pedantic
SuperPipe operators under one rule and then to enumerate all of the
syntatic shortcuts under the leaner LeanOp rule (which used to be
Operation).  With this layout, it's really clear where the lean grammar
diverges from the pedantic grammar.  We can subsequently use this
pattern to make a lean SuperSQL alongside the pedantic SuperSQL.

We also deleted the Cast rule from the LeanOp location as it
was redundant with the Expr rule for implied yield.  This undoes the
change in #3697 but retains the test, which still passes.