refactor api between compiler and parser (#5401)
This commit clears up the interface between the parser and compiler
and separates the collating of errors to source files into a separate
package that is shared by the parser and semantic pass.  This gets
rid of the awkward error handling that required binding the source-file
set back to the error list with a complex type assertion when
handling what should just be a normal Go error.  The only time this binding
trick now must happen is in the API when the client library gets a response
form the server. (We could further clean things up here by pushing this
model through the service API but we don't do this in this commit.)