make aggregate functions and expression functions case insensitive (#5404)
This commit changes the semantic pass (and PEG parser for a couple
cases) to accept mixed-case names for predefined functions and
aggregate functions.  User-defined functions remain case sensitive.
The invariant is that the transformations are all handled in the
semantic pass on the DAG presumes everything relevant here has been
converted to lower case.