Remove optimizer.Optimizer.concurrentPath needMerge return param (#5411)
The needMerge return parameter is redundant because it should always be
equal to len(outputKeys)==0.  (Intuitively, if len(outputKeys)>0, then
the output is ordered and a dag.Merge is necessary, but if
len(outputKeys)==0, then the output is unordered and a dag.Combine
suffices.)

Note that concurrentPath can currently return needMerge==true when
len(outputKeys)==0 but that behavior is erroneous because a merge
requires a merge key.