add support for "*" in sql select (#5413)
This commit adds super for "*" in select clauses as an
empty ast.Assignment.  There isn't yet support for id.*
style selections. We also reworked the selection naming
convention to use the last name in a dotted path, which
is what SQL does.

The SQL semantic logic was also reorged and renamed a bit
to better position it for the comming work on scoping and
binding.