add tuple expressions (#5428)
This commit adds a tuple expression to the query language.  Since the
data model and runtime do not support native tuplesm, these expressions
are mapped onto records with columns c0, c1, ...  Datafusion does this
and other SQLs map tuples to structs with unnamed fields.

This will be needed by a subsequent PR implementing the VALUES clause.