Implement vector join operator (#5659)
This is a simple hash join with lots of room for improvement.  Because
it must read its entire right input before reading any values from its
left input, it can deadlock on self joins.

For deterministic output, it requires the addition of an output sort in
tests that don't already have one.

(runtime/ztests/parallel.yaml is removed as it duplicates other tests.)