small refactor to SQL name resolution (#5675)
This commit refactors the logic that performs name resolution for
SQL names so that path concatenation is separated from the table/column
lookups making the logic a bit more clear.  We also changed the return
value of resolve() to an expression instead of path to make it more general.
This will be needed by a forthcoming PR providing expression access
to the "this" concept within SQL scopes.