Fix data race in vector Parquet reader with pruner (#5729)
In a vector Parquet reader with a pruner, multiple goroutines may use
the pruner concurrently, but the pruner is a sam/expr.Evalautor, some
implementations of which (notably sam/expr.DotExpr) are unsafe for
concurrent use.  Fix this by changing the pruner paramter to a
zbuf.Filter for runtime/exec.Environment.VectorOpen,
parquetio.NewVectorReader, and vngio.NewVectorReader.