remove debuging info
This commit is almost working.  There's some bugs and this is hard
to debug because of the custom unmarshaling and it will be hard to
maintain because of the maintenance of the shadow definitions that
mirror csup.MetaData.  A better approach is to simplify the serialization
with a flat table that we'll call csup.Context.  We are going to pursue
this approach and mark this commit as csup-shadow-save.