unify pushdown interface (#5775)
This commit unifies the pushdown logic for filtering, BSUP filtering,
CSUP metadata filtering, and projection into a single interface
called zbuf.Pushdown replacing zbuf.Filter.  We also added projections
to the data filters and arranged for the CSUP reader to project
only the needed metadata values according to this projection.

These changes are needed to an upcoming PR that will improve the
efficiency of CSUP by doing lazy type loading.