refactor null bit vectors and vector.Bool (#5796)
This commit separates the implementation of bit vectors from
the vector.Bool runtime vector.  We were using vector.Bool as
a bit vector even when it wasn't a runtime vector causing confusion
in the code (my bad for doing this originally).  We have moved the
bit vector logic from vector.Bool into a new package in vector/bitvec
and replaced all the vector.Any.Null bitmaps with a bitvec.Bits.