lazy vectors
This commit adds lazy vectors to CSUP and the vector runtime.
This is accomplished by adding new methods to vectors that
return values, nulls, etc and when the underlying data isn't
present, it is loaded on demand.  Behavior is unchanged for vectors
created with the previous vector.New* API.

This approach improve performance for queries where the optimizer
resorts to conservative projections due to unpredictability, where
only the data actually needed by the query is loaded.  The projection
is still important for pruning the amount of vcache shadow data that
must be decoded when there are many types present in the input data.
A future improvement could be to incrementally load shadow vectors
based on demand too.