Change top operator to behave like "sort | head" (#5804)
Change the top operator to accept the same options as the sort operator
(plus a limit expression) and behave exactly like "sort | head" so that,
e.g., "top -nulls first -r 5 a, b asc, c desc" produces the same output
as "sort -nulls first -r a, b asc, c desc | head 5".

This is groundwork for an optimization that replaces "sort | head" and
"sort | tail" with top.