Attempt to fix TestAggregateStreamingSpill flakiness (#5906)
This test fails occasionally on Windows in CI like this:

    --- FAIL: TestAggregateStreamingSpill (0.35s)
	aggregate_test.go:118:
		    Error Trace:	D:/a/super/super/runtime/sam/op/aggregate/aggregate_test.go:118
							    D:/a/super/super/runtime/sam/op/aggregate/aggregate_test.go:68
							    D:/a/super/super/zbuf/batch.go:41
							    D:/a/super/super/zbuf/batch.go:169
							    D:/a/super/super/runtime/sam/op/aggregate/aggregate_test.go:127
							    D:/a/super/super/runtime/sam/op/aggregate/aggregate_test.go:135
		    Error:      	"200" is not less than "200"
		    Test:       	TestAggregateStreamingSpill
    FAIL
    FAIL	github.com/brimdata/super/runtime/sam/op/aggregate	0.729s

One possible explanation is that the query consumes values from the
countReader in batches that may be too large to produce streaming output
given the number of input values.  Address that by changing countReader
to implement zbuf.ScannerAble so it can feed single-value batches to the
query.