join: Lift parent sequence into Fork path (#5954)
This commit changes the behavior of joins with right inputs so that the
semantic analyzer lifts the parent flowgraph into the created Fork operator
instead of using the Pass operator. This corrects a bug when using
yields on right side of join that essentially causes the join to act like
a cross join.

It also changes the behavior of semantic.HasSource so that a program
starting with a Fork operator and containing a mix of Sourced and
non-Sourced paths will get the default or null scanner added to it.