add double-colon style casts (#5994)
This commit adds colon-style casts to the language a la PostgreSQL.  
The syntax is <expr>::<expr> where the righthand side should evaluate 
to a type, or <expr>::<type> where the <type> can be any SuperSQL type
aside from named types (as they otherwise collide with identifiers).

Currently, quoted strings are allowed as the type arg but this is a feature that
we will likely remove from the underlying cast function.