change SUP type decorators to use double-colon casts (#6009)
This commit changes the SUP format to use double-colon casts instead
of parentheses so that SUP values are syntax-compatible with SuperSQL
literals (except for integer typedefs).  This required a change to
the top-level type syntax for unions and typedefs where these entities
must be parenthesized when they appear as a top-level decorator.
It is likely possible to elide these parentheses inside of values
but we would need to pass down such context and do a bit more
work in the formatter and parser.  We can do this later (or never).

We also added syntax to SuperSQL to handle ::=typedefs.

Comprehensive documentation updates are to follow in the book branch.