cleanup the enum type (#6035)
This commit cleans up the enum type and brings it into alignment
with other SQLs.  In this new approach, an enum is represented as
a string cast to an enum type.  We updated the SUP parser and formatter,
added enum types to the SuperSQL grammar (which weren't previously
implemented), and adapted the cast operator in the runtime to handle casts
of strings to enum values (and named enums).

Doc updates for SUP and SuperSQL will appear in the forthcoming book PRs.