Rework UDFs (#6044)
This commit changes how udfs work so that they no longer rely on
variables but instead rewrite the context value as a record with a field
for every function parameter paired with its call expression. It also
removes UDFs dependency on expr.Context to keep track of current stack
size of recursive function calls.