simplify stack depth limit checking in expr Eval (#6053)
This commit simplifies the logic in the stack depth calculation
by keeping the depth state in each call site instance instead of
trying to share it.  So instead of tracking a global depth, we
just compute a local depth for each function and halt recursion
with an error when the limit is locally exceeded.