use file extension to imply format and avoid auto-detection (#6063)
This commit changes the heuristic for detecting input file (or URL)
formats by first checking for known file extensions when the input
path is known.  Only when this heuristic fails is autodetection
employed.  This was already being done for parquet and csup files
(not URLs) and this commit generalizes the approach to all support
file types and includes the check for URLs.

When files are gzipped, this test fails and the autodetection logic
detects the gzip and the enclosing format.  We might want to extend
this approach here to gzipped files in the future, e.g., where
".json.gz" implies gzipped JSON without having to do autodetection.