unify and simplify field assignments (#6072)
This commit makes field assignments and the optionality of
"name:=..." a consistent element of the SuperSQL syntax. This
means the put operator now works this way. We updated the
semantic pass to always derive a name for all missing LHS
using zfmt.  Otherwise the semantic logic that analyzed
assignments in various operators was already handling the general
case of a missing/null LHS so there were no changes needed here.