simplify regexp/grep syntax (#6115)
This commit cleans up the regexp/grep syntax by isolating the
regexp pattern and glob syntax to search expressions and using
conventional strings for the functions regexp(), regexp_replace(),
and grep().  This allowed us to remove a bunch of special case
syntax for these functions from the PEG grammar.

This change means that globbing is no longer available for grep,
which now uses only regular expressions.  If we want to allow
globbing, we can enable that with a function that converts glob
strings to regexps or perhaps introduce a g"glob" style string.

We also removed the implied "this" argument from grep and now
require two arguments.

Detailed docs updates are forthcoming in the book branch.