simplify text entity syntax (#6122)
This commit simplifies the syntax rules for text entites by
creating a new grammar rule for a "simple URL".  This makes
usage and documentation clearer.

While here, we changed the ":meta" meta-modifier syntax to a
"name" (the same rules as field names), where the it can be
anything if quoted and otherwise must have identifier syntax.

We also eliminated the unused PEG rule PoolAt and simplified
the rule for KSUID as it's now used in just one place (KSUID
overlaps with TextChars so it's not needed in the Text rule).

Documentation changes are forthcoming in the book branch.