compiler/optimizer: fix missing field bug in liftFilterOps, mergeValuesOps (#6219)
Both optimizations do the wrong thing when they encounter a field known
to be missing.  For example, liftFilterOps rewrites
`values {a} | where b==1` to `where b==1 | values {a}` instead of
`where error("missing")==1 | values {a}`, and mergeValuesOps rewrites
`values {a} | values b` to `values b` instead of `values error("missing")`.