automatically promote a user-op in an expression to a subquery (#6226)
This commit changes the handling of user-ops referenced in expressions
from an error to a subquery.  When the user-op definition begins with a from,
then the subquery is spliced into the expression as an uncorrelated query.
Otherwise, it is dependent on its input and the user-op is expanded
into the pattern (values this | userOp(params)).  User-ops references can also
be referenced inside array expression to get the automatically appended
"collect(this)".