remove scopes from dag and simplify functions, fix scoping bugs (#6233)
This commit removes dag.Scope and simplifies function handling
in the DAG by having a global table of functions indexed by a tag.
There is no need for scoping in the DAG since lexical scoping is
all resolved by the end of the semantic pass.

This undoes the complexity of the recent dag.FuncRef addition,
which took things in the wrong direction (my bad).

All lambda functions are now compiled alongside named user functions
in the DAG function table.  This sets us up to handle passing lambdas
to functions (in addition to ops), but this will come in a subsequent PR.

This commit also fixes some scoping problems where the lookup-by-name
approach for function resolution didn't work quite right.  See new tests
in compiler/semantic/ztests.