support for restartable dag.FileScans so repeated from-subqueries work (#6251)
This commit adds support for restarting file scans so that repeated
from-subqueries involving file paths work. This is a temporary
solution to make progress on docs and testing.  It does not work
for databases and fails silently for non-seekable streams.

A more comprehensive solution will come in a future PR.

To make this work, we made some changes to the hashjoin operator
since it wasn't correctly following the done protocol.  Here, we wrap
the bufPuller with another done bool and implement strict adherence
to the done protocol so Pull can now be called multiple times after
the parent is done without restarting the parent.  This is ok because
the top hashjoin object will reallocate the entire entity on restart and
with this new logic, the join parents should experience correct behavior
 and restart on the realloc.