add let statements for query references (#6252)
This commit adds support for named queries with a "let" syntax.
After binding a name to a query, the query may be invoked using
"from" or in an expresseion, which implies an expression subquery.

We also fixed a problem where user function and operator names
would collide with field name references in expresions and l-values.

Docs updates coming soon from the book branch.

Closes #6248