SQL select star static schema column names (#6264)
This commit changes the generated values operator dag in a SQL query that
uses select * so that if the underlying schema is static all the selected
columns will be enumerated instead of using record spread expressions. This
change allows us to optimize queries that could otherwise not be optimized
since the source of a field is obfuscated when spreads are used.