unify SQL and pipe name inference and simplify single by-only agg output (#6263)
This commit unifies the way l-values are inferred for expressions
that omit the LHS of a field assignment or the AS clause in a select
SELECT expression.  ANSI SQL does not have specific guidance on how
to infer arbitrary expressions so we blended in some of the existing
logic from the pipe syntax.

We also changed how "this" is handled.  Instead of raising an error,
the l-value becomes "that".  We also adapted the heuristic for a single
column agg with no groupings to work also for a single grouping with
no aggs (i.e., emitting the values instead of a single-column record).
These two changes  means that "by this" now works and behaves like distinct.

Overall, this improves the consistency and orthogonality of the language.

Per our discussion, I tried to change cut to take path expressions
instead of assignments but this had an impact on the optimizer tests.
When changing cut to values in these tests, they were no longer
optimized the same way, which means the optimizer needs some work.
The optimizer needed this work anyway, but rather than do this here, I backed out
some of the cut changes and will defer the updates to a subsequent PR.