add semantic tree (#6267)
This commit adds a new data structure that represents the query
tree in between the AST and the DAG.  This will set us up to
bring type checking to dynamic data using the semantic tree.

Each entity in the semantic tree has an exported ast.Node interface
pointer so everything implements ast.Node and locators still work
(unlike in the DAG).

We also fixed how consts are evaluated and how EvalAtCompileTime
is invoked.  In this new approach, we re-enter the semantic
analysis for any constant function needing eval and have a
"maybe" version that ignores errors and a "must" version that
captures errors.  Functions and subqueries are now properly
handled when evaluating consts.  We also walk the sem tree to
figure out which expressions are const instead of relying on
the unreliable error("missing") hack that we previously used.

Expression matching for SQL group-by expressions is now done
on the sem tree (instead of DAG) and utilizes some deep-equal
logic that traverses the tree.

This is some formatting logic for sem tree in compiler/semantic/fmt.go,
which is useful for printf debugging and will be tied into
"super compile" in a subsequent PR.

There are still some problems with declaration dependencies
(e..g, fn decls that depend on const decls in same block and
vice versa) and we added some skipped tests for this.  We will
fix this in an upcoming branch decl-dep.

Type checking will come in a subsequent PR.