improve declaration bindings with lazy evaluation (#6268)
This commit improves how declaration bindins work in a given scope
by binding all the symbols as encountered then resolving declaration
bodies on demand.  This means that dependencies are handles in an
as-used basis and the order of declaration does not matter (for a
given scope block).

This is needed by the forthcoming type checker.