unify naming in package dag (#6279)
This commit cleans up the naming conventions in package dag, remove some
unused code, and reorders some code to conform with alphabetical order
of type names.  All dag.Ops now have a suffix of "Op" or "Scan" and all
dag.Exprs have a suffix of "Expr".

This is strictly a naming/style change and no logic is changed or
introduced here.