remove crop(), fill(), fit(), order(), and shape() (#6286)
cast() is generally sufficient though we may add a variant that doesn't
add missing fields with a null value when casting records.

runtime/sam/expr.ConstShaper has to stick around because it's used in
runtime/sam/op/fuse.  We can remove it in favor of
runtime/sam/expr/function.cast after we enhance that latter to handle
union-typed record fields in the target type.