unify AST naming (#6288)
This commit unifies the naming patterns of the AST to be consistent
with the DAG and the semantic tree.  Ops have suffixes "Op" and
ast.Exprs have suffix "Expr".

We removed the Kind field from structs that aren't part of sum types.
We also separate ast.ID into ID and IDExpr to differentiate the
use of ID in an expression versus identifiers in other non-expression
entities.

There are further improvements that can be made but this is a big step
in the right direction.