add table references and "this" to SQL expressions (#6292)
This commit adds support for referring to tables in a SQL expression,
resulting in a record representing the table row (as in duckdb and
somewhat similarly in postgres).  We also added support for referencing
"this" in a SQL expression, which refers to the input relation
in SELECT and WHERE clauses, the output relation in HAVING clauses,
and the input relation for arguments (and WHERE clauses) of agg functions
appearing in HAVING clauses.

This new logic causes an error for table references of dynamic schemas.
This is to avoid a situation where "select T from T" refers to the table
in a dynamic schema, then when a schema shows up for that same data, the
query compiles differently to a field reference of T inside T (following
postgres and duckdb scope precedence).  When it is desired to refer to
the table of a dynamic source, the special value "this" can be used instead.
In general, query semantics should be identical whether types/schemas are known
or unknown; if this isn't the case anywhere here, then it's a design bug.

We also added an escape valve for referring to a SQL column named "this",
which is simply denoted with double quotes.

Finally, these changes exposed a problem in the as-name inference algorithm,
where the internal DAG paths would show up, so we updated the inference code
to strictly use the AST instead of a mix of the AST and sem tree.