runtime/vam/op: fix combine EOS bug (#6317)
In a call to runtime/vam/op.Combine.Pull, when all parents are observed
to be blocked, they are immediately unblocked and EOS is returned.
Unblocking there is premature since new results from parents won't be
consumed unless Pull is called again.  Fix this by waiting to unblock
parents until the next call to Pull.