add support for re-entering subqueries in recursive functions (#6325)
This commit addresses an unfinished aspect of subqueries when they
are invoked from recursive functions.  Previously, the subquery would
incorrectly block as Eval() is called concurrently violating our invariant.

The fix is to keep a stack of non-shared subquery instances that can
be used when recursion is detected.  This involves calling the rungen
builder from the runtime whenever we need a new instance, which turned
out to be surprisingly easy to implement.