simplify ::-cast and cast-as to use only types not expressions (#6329)
This commit simplifies the :: and SQL "cast as" syntax to allow
only types (not expressions that result in type values) on the RHS.
We wanted to get rid of the cast() function but are retaining it for
casting to types resulting from expressions.  This is because we
could not resolve the ambiguity between arbitrary parenthesized
types with type names and generic expressions with names as field
references.  These changes will simplify the docs regarding casts
that are coming in a subsequent PR.