rework handling of command line input files (#6334)
Input files appearing on the command line are fed to the compiler as a
single sio.Reader, preventing the semantic analyzer from seeing their
super.Types.  Feed them in as paths instead.  (The semantic analyzer
does not take advantage of this yet.)

As a side effect, dag.FileScan is now restartable in the vector runtime
so it can now execute repeated "from <file>" subqueries.