Add schema support for SQL set operations (#6356)
This commit changes the behavior of SQL set operations in that it
requires both the left and right side to have static schemas and for
both sides to have an equal number of columns otherwise errors are
returned. The right side record fields are renamed to that of the left
side if the column names are not the same. A valid schema is returned so
that two set operations can be joined.

This commit also fixes an union with SQLUnion where the parsed AST Loc
was not set and adds the Node interface to ast.SQLQueryBody so errors
can be reported on these elements.