cleanly separate pipe from and SQL FROM (#6405)
This commit cleanly separates the syntax of the pipe-operator "from"
and the SQL FROM clause.  This means that table aliases are no longer
allowed with pipe from (even though they were ignored).  Also, a
relational FROM is no longer possible without a SELECT clause
(required a "SELECT *" to do a straight relational join).  We can
revisit this if there's interest as the grammar should be able
to handle it.

This better aligns the PEG grammar and semantic pass with the
actual SQL syntax and will make it easier to document the
FROM clause and the "from" pipe operator.

Documentation is forthcoming from the book-sql branch.