change "from eval(...)" to "from f'{...}' (#6450)
This commit changes the "robot from" syntax from the eval() structure
to a simple f-string.  This recognizes that entities in from operators
are always identified as strings.  The ability to scan the entities in
an array of strings is no longer available with this change, but this can
always be emulated by doing an "over" upstream of of the "robot from"
on the array of strings.

Docs updates are coming in a subsequent PR from the book-sql branch.